options: Allow users to pick sites to target
This commit allows users to specify which sites they want to target with the
extension. We use an embedded option page.

Note, however, that this introduces an important usability problem with the
current implementation: We previsouly hid the body of the page in CSS to avoid
the page showing up briefly before the popup loads. However, now that we apply
the extension to all URLs, it is not acceptable that we hide the body
automatically - this would "slow down" all websites the user visits. We
therefore no longer hide the body by default. This means that the users again
get to see the "blocked" website before the URL loads. This is not ideal, but a
better temporary solution than the alternative.